requirements.txt
python-binance>=1.0.17
python-dotenv>=1.0.0
loguru>=0.7.2
pandas>=2.0.0
numpy>=1.24.0